pyproject.toml
[tool.poetry]
name = "thegrandlocus"
version = "0.1.0"
description = "Modernized version of The Grand Locus blog"
authors = ["Guillaume Filion <guillaume.filion@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.9"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
google-cloud-datastore = "*"
google-cloud-storage = "*"
Jinja2 = "*"
markdown = "*"
beautifulsoup4 = "*"
python-multipart = "*"
Pydantic = "*"
Authlib = "*"
itsdangerous = "*"
httpx = "*"
google-auth-oauthlib = "*"
pydantic-settings = "*"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src",
  "models_",
]
testpaths = [
    "tests",
]
markers = [
    "diskspace: marks tests as requiring disk space (deselect with '-m \"not diskspace\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
